Only remove symlinks to vendor dirs on non-Windows OSes
On Windows, we're unable to distinguish between symlinks and junctions. Deleting
junctions is dangerous and appears to be able to delete the underlying folder,
not just the junction that's pointing to a folder. That's unacceptably
dangerous: a malicious repo could contain a junction named 'vendor' pointing at
'C:\\' and we'd delete all the user's data.
