+++ result_test.go
@@ -140,197 +140,3 @@ func BenchmarkCreateVendorTree(b *testing.B) {
 	sm.Release()
 	os.RemoveAll(tmp) // comment this to leave temp dir behind for inspection
 }
-
-func TestStripVendor(t *testing.T) {
-	t.Run("vendor directory", stripVendorTestCase(fsTestCase{
-		before: filesystemState{
-			dirs: []fsPath{
-				fsPath{"package"},
-				fsPath{"package", "vendor"},
-			},
-		},
-		after: filesystemState{
-			dirs: []fsPath{
-				fsPath{"package"},
-			},
-		},
-	}))
-
-	t.Run("vendor file", stripVendorTestCase(fsTestCase{
-		before: filesystemState{
-			dirs: []fsPath{
-				fsPath{"package"},
-			},
-			files: []fsPath{
-				fsPath{"package", "vendor"},
-			},
-		},
-		after: filesystemState{
-			dirs: []fsPath{
-				fsPath{"package"},
-			},
-			files: []fsPath{
-				fsPath{"package", "vendor"},
-			},
-		},
-	}))
-
-	t.Run("vendor symlink", stripVendorTestCase(fsTestCase{
-		before: filesystemState{
-			dirs: []fsPath{
-				fsPath{"package"},
-				fsPath{"package", "_vendor"},
-			},
-			links: []fsLink{
-				fsLink{
-					path: fsPath{"package", "vendor"},
-					to:   "_vendor",
-				},
-			},
-		},
-		after: filesystemState{
-			dirs: []fsPath{
-				fsPath{"package"},
-				fsPath{"package", "_vendor"},
-			},
-		},
-	}))
-
-	t.Run("nonvendor symlink", stripVendorTestCase(fsTestCase{
-		before: filesystemState{
-			dirs: []fsPath{
-				fsPath{"package"},
-				fsPath{"package", "_vendor"},
-			},
-			links: []fsLink{
-				fsLink{
-					path: fsPath{"package", "link"},
-					to:   "_vendor",
-				},
-			},
-		},
-		after: filesystemState{
-			dirs: []fsPath{
-				fsPath{"package"},
-				fsPath{"package", "_vendor"},
-			},
-			links: []fsLink{
-				fsLink{
-					path: fsPath{"package", "link"},
-					to:   "_vendor",
-				},
-			},
-		},
-	}))
-
-	t.Run("vendor symlink to file", stripVendorTestCase(fsTestCase{
-		before: filesystemState{
-			files: []fsPath{
-				fsPath{"file"},
-			},
-			links: []fsLink{
-				fsLink{
-					path: fsPath{"vendor"},
-					to:   "file",
-				},
-			},
-		},
-		after: filesystemState{
-			files: []fsPath{
-				fsPath{"file"},
-			},
-			links: []fsLink{
-				fsLink{
-					path: fsPath{"vendor"},
-					to:   "file",
-				},
-			},
-		},
-	}))
-
-	t.Run("chained symlinks", stripVendorTestCase(fsTestCase{
-		before: filesystemState{
-			dirs: []fsPath{
-				fsPath{"_vendor"},
-			},
-			links: []fsLink{
-				fsLink{
-					path: fsPath{"vendor"},
-					to:   "vendor2",
-				},
-				fsLink{
-					path: fsPath{"vendor2"},
-					to:   "_vendor",
-				},
-			},
-		},
-		after: filesystemState{
-			dirs: []fsPath{
-				fsPath{"_vendor"},
-			},
-			links: []fsLink{
-				fsLink{
-					path: fsPath{"vendor2"},
-					to:   "_vendor",
-				},
-			},
-		},
-	}))
-
-	t.Run("circular symlinks", stripVendorTestCase(fsTestCase{
-		before: filesystemState{
-			dirs: []fsPath{
-				fsPath{"package"},
-			},
-			links: []fsLink{
-				fsLink{
-					path: fsPath{"package", "link1"},
-					to:   "link2",
-				},
-				fsLink{
-					path: fsPath{"package", "link2"},
-					to:   "link1",
-				},
-			},
-		},
-		after: filesystemState{
-			dirs: []fsPath{
-				fsPath{"package"},
-			},
-			links: []fsLink{
-				fsLink{
-					path: fsPath{"package", "link1"},
-					to:   "link2",
-				},
-				fsLink{
-					path: fsPath{"package", "link2"},
-					to:   "link1",
-				},
-			},
-		},
-	}))
-}
-
-func stripVendorTestCase(tc fsTestCase) func(*testing.T) {
-	return func(t *testing.T) {
-		tempDir, err := ioutil.TempDir("", "TestStripVendor")
-		if err != nil {
-			t.Fatalf("ioutil.TempDir err=%q", err)
-		}
-		defer func() {
-			if err := os.RemoveAll(tempDir); err != nil {
-				t.Errorf("os.RemoveAll(%q) err=%q", tempDir, err)
-			}
-		}()
-		tc.before.root = tempDir
-		tc.after.root = tempDir
-
-		tc.before.setup(t)
-
-		if err := filepath.Walk(tempDir, stripVendor); err != nil {
-			t.Errorf("filepath.Walk err=%q", err)
-		}
-
-		tc.after.assert(t)
-	}
-}
